backend/requirements.txt
flask
pymongo
flask-cors
bcrypt
python-dotenv